Edit function working!
